+++ src/js/album.js
@@ -1,0 +1,7 @@
+/* global $:true */
+
++ function($) {
+  "use strict";
+
+
+}($);