+++ changelog.md
@@ -1,3 +1,20 @@
+# V0.6.0
+
+对应 `weui` 版本为 `0.4.0`
+
+- 新增组件 `Select`、`Popup`
+- 修复了 `Picker` 和 `Canlendar` 在部分安卓手机上无法显示的bug（原因是安卓 4.4 以下不支持 `calc`）。
+- 修复了 `City Picker` 中设置初始值之后可能导致的bug。
+- `City Picker` 增加了一个 `showDistrict` 配置，可以配置是否选择 `区`。
+- 修复了 `Picker` 在 iOS 系统中无法出现弹出动画的bug。
+- 现在打开 `Picker` 时会自动关闭上一个打开的 `Picker`
+- 修复了使用 `weui.min.css` 时由于 `z-index` 压缩问题导致对话框被遮住的Bug。
+- 更新了 `Search Bar` 的逻辑，现在`取消`,`清除`等按钮更符合逻辑。
+- 对话框增加了一个 `autoClose` 配置，可以配置点击按钮之后是否自动关闭对话框。
+- `Datetime Picker` 中可以设置时间范围了。
+- 更新了 demos 中的图标
+- 在JS和CSS文件中加入了版本号等说明
+
 # v0.5.1 - 2016/03/24
 
 对应 `weui` 版本为 `0.4.0`